Choosing Affirmations Alert is its own componenet
