+++ HigherSelf/app/modals/practiceAffirmations.tsx
@@ -1,14 +1,13 @@
+import CustomChooseAffirmationAlert from '@/components/CustomChooseAffirmationAlert';
 import { useSavedAffirmations } from '@/context/SavedAffirmationContext';
 import { STORAGE_KEYS } from '@/data/HigherSelf_StorageKeys';
 import type { Affirmation } from '@/types/affirmations';
 import { Ionicons, MaterialCommunityIcons } from '@expo/vector-icons';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { router } from 'expo-router';
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import {
   ActivityIndicator,
-  FlatList,
-  Modal,
   Pressable,
   ScrollView,
   StyleSheet,
@@ -17,7 +16,6 @@ import {
 } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
-type PracticeTab = 'saved' | 'custom';
 type PracticeStep = 'queue' | 'recording';
 
 type StoredAffirmationLike =
@@ -28,11 +26,6 @@ type StoredAffirmationLike =
       affirmation?: string;
       category?: string;
     };
-
-const PRACTICE_TABS: { id: PracticeTab; label: string }[] = [
-  { id: 'saved', label: 'Saved' },
-  { id: 'custom', label: 'My Own' },
-];
 
 const getTodayKey = () => new Date().toISOString().slice(0, 10);
 
@@ -92,7 +85,6 @@ export default function PracticeAffirmationsScreen() {
   >([]);
   const [customAffirmations, setCustomAffirmations] = useState<Affirmation[]>([]);
   const [queuedAffirmations, setQueuedAffirmations] = useState<Affirmation[]>([]);
-  const [activeTab, setActiveTab] = useState<PracticeTab>('saved');
   const [showPicker, setShowPicker] = useState(false);
   const [screenStep, setScreenStep] = useState<PracticeStep>('queue');
   const [isBootstrapping, setIsBootstrapping] = useState(true);
@@ -146,7 +138,7 @@ export default function PracticeAffirmationsScreen() {
   }, []);
 
   useEffect(() => {
-    if (activeTab !== 'saved') {
+    if (!showPicker) {
       return;
     }
 
@@ -170,7 +162,7 @@ export default function PracticeAffirmationsScreen() {
     };
 
     loadSavedAffirmationsFromDevice();
-  }, [activeTab]);
+  }, [showPicker]);
 
   useEffect(() => {
     if (isBootstrapping) {
@@ -191,20 +183,6 @@ export default function PracticeAffirmationsScreen() {
     persistQueue();
   }, [isBootstrapping, queuedAffirmations]);
 
-  const affirmationsByTab = useMemo(
-    () => ({
-      saved: deviceSavedAffirmations,
-      custom: customAffirmations,
-    }),
-    [customAffirmations, deviceSavedAffirmations]
-  );
-
-  const visibleAffirmations = affirmationsByTab[activeTab];
-  const queuedIds = useMemo(
-    () => new Set(queuedAffirmations.map((affirmation) => affirmation.id)),
-    [queuedAffirmations]
-  );
-
   const toggleQueuedAffirmation = (affirmation: Affirmation) => {
     setQueuedAffirmations((currentQueue) => {
       if (currentQueue.some((item) => item.id === affirmation.id)) {
@@ -222,39 +200,6 @@ export default function PracticeAffirmationsScreen() {
 
     setShowPicker(false);
     setScreenStep('recording');
-  };
-
-  const renderAffirmationCard = ({ item }: { item: Affirmation }) => {
-    const selected = queuedIds.has(item.id);
-
-    return (
-      <View
-        style={[
-          styles.affirmationCard,
-          selected && styles.affirmationCardSelected,
-        ]}
-      >
-        <Text style={styles.affirmationCardText}>{item.text}</Text>
-
-        <Pressable
-          accessibilityLabel={
-            selected ? 'Remove affirmation from queue' : 'Add affirmation to queue'
-          }
-          onPress={() => toggleQueuedAffirmation(item)}
-          style={({ pressed }) => [
-            styles.addButton,
-            selected && styles.addButtonSelected,
-            pressed && styles.addButtonPressed,
-          ]}
-        >
-          <Ionicons
-            color="#F5F7FA"
-            name={selected ? 'checkmark' : 'add'}
-            size={22}
-          />
-        </Pressable>
-      </View>
-    );
   };
 
   const isLoading = savedLoading || isBootstrapping;
@@ -363,106 +308,15 @@ export default function PracticeAffirmationsScreen() {
           )}
         </View>
 
-        <Modal
-          animationType="fade"
-          onRequestClose={() => setShowPicker(false)}
-          transparent
+        <CustomChooseAffirmationAlert
+          customAffirmations={customAffirmations}
+          onClose={() => setShowPicker(false)}
+          onGoToRecording={openRecordingScreen}
+          onToggleAffirmation={toggleQueuedAffirmation}
+          queuedAffirmations={queuedAffirmations}
+          savedAffirmations={deviceSavedAffirmations}
           visible={showPicker}
-        >
-          <View style={styles.modalBackdrop}>
-            <View style={styles.modalCard}>
-              <View style={styles.modalHeader}>
-                <View>
-                  <Text style={styles.modalTitle}>Choose Affirmations</Text>
-                  <Text style={styles.modalSubtitle}>
-                    Add the affirmations you want to recite today.
-                  </Text>
-                </View>
-
-                <Pressable
-                  onPress={() => setShowPicker(false)}
-                  style={styles.closeButton}
-                >
-                  <Ionicons color="#F5F7FA" name="close" size={22} />
-                </Pressable>
-              </View>
-
-              <View style={styles.tabRow}>
-                <View style={styles.tabSwitcher}>
-                  {PRACTICE_TABS.map((tab) => {
-                    const isActive = tab.id === activeTab;
-
-                    return (
-                      <Pressable
-                        key={tab.id}
-                        onPress={() => setActiveTab(tab.id)}
-                        style={[styles.tabButton, isActive && styles.tabButtonActive]}
-                      >
-                        <Text
-                          style={[styles.tabButtonText, isActive && styles.tabButtonTextActive]}
-                        >
-                          {tab.label}
-                        </Text>
-                      </Pressable>
-                    );
-                  })}
-                </View>
-              </View>
-
-              <Text style={styles.tabMetaText}>
-                {affirmationsByTab[activeTab].length} affirmation
-                {affirmationsByTab[activeTab].length === 1 ? '' : 's'}
-              </Text>
-
-              <View style={styles.listRegion}>
-                {visibleAffirmations.length ? (
-                  <FlatList
-                    contentContainerStyle={styles.affirmationListContent}
-                    data={visibleAffirmations}
-                    keyExtractor={(item) => item.id}
-                    renderItem={renderAffirmationCard}
-                    ItemSeparatorComponent={() => <View style={styles.listSeparator} />}
-                    showsVerticalScrollIndicator={false}
-                    style={styles.affirmationList}
-                  />
-                ) : (
-                  <View style={styles.emptyTabState}>
-                    <Text style={styles.emptyTabTitle}>
-                      {activeTab === 'saved'
-                        ? 'No saved affirmations yet'
-                        : 'No custom affirmations yet'}
-                    </Text>
-                    <Text style={styles.emptyTabCopy}>
-                      {activeTab === 'saved'
-                        ? 'Liked affirmations from the home screen will appear here.'
-                        : 'Affirmations you create yourself will appear in this tab.'}
-                    </Text>
-                  </View>
-                )}
-              </View>
-
-              <View style={styles.modalFooter}>
-                <View style={styles.queueCountPill}>
-                  <Text style={styles.queueCountText}>
-                    {queuedAffirmations.length} Affirmation
-                    {queuedAffirmations.length === 1 ? '' : 's'} Ready
-                  </Text>
-                </View>
-
-                <Pressable
-                  disabled={!queuedAffirmations.length}
-                  onPress={openRecordingScreen}
-                  style={[
-                    styles.primaryButton,
-                    !queuedAffirmations.length && styles.primaryButtonDisabled,
-                  ]}
-                >
-                  <Text style={styles.primaryButtonText}>Go To Recording</Text>
-                </Pressable>
-              </View>
-            </View>
-          </View>
-        </Modal>
+        />
       </SafeAreaView>
     </View>
   );
@@ -676,179 +530,5 @@ const styles = StyleSheet.create({
     fontSize: 15,
     lineHeight: 23,
     fontWeight: '500',
-  },
-  modalBackdrop: {
-    flex: 1,
-    backgroundColor: 'rgba(3, 7, 18, 0.7)',
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingHorizontal: 18,
-  },
-  modalCard: {
-    width: '100%',
-    maxWidth: 430,
-    height: '86%',
-    borderRadius: 28,
-    backgroundColor: '#101726',
-    borderWidth: 1,
-    borderColor: 'rgba(255,255,255,0.08)',
-    paddingHorizontal: 18,
-    paddingTop: 18,
-    paddingBottom: 20,
-  },
-  modalHeader: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'flex-start',
-    gap: 16,
-  },
-  modalTitle: {
-    color: '#F5F7FA',
-    fontSize: 24,
-    fontWeight: '700',
-  },
-  modalSubtitle: {
-    color: '#C6CDD8',
-    fontSize: 14,
-    lineHeight: 20,
-    marginTop: 6,
-    maxWidth: 260,
-  },
-  tabRow: {
-    marginTop: 22,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  tabSwitcher: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: 'rgba(255,255,255,0.04)',
-    borderWidth: 1,
-    borderColor: 'rgba(255,255,255,0.06)',
-    borderRadius: 18,
-    padding: 4,
-  },
-  tabButton: {
-    minWidth: 116,
-    borderRadius: 14,
-    paddingVertical: 12,
-    paddingHorizontal: 20,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  tabButtonActive: {
-    backgroundColor: 'rgba(139, 92, 246, 0.22)',
-  },
-  tabButtonText: {
-    color: '#C6CDD8',
-    fontSize: 15,
-    fontWeight: '700',
-  },
-  tabButtonTextActive: {
-    color: '#F5F7FA',
-  },
-  tabMetaText: {
-    color: '#94A3B8',
-    fontSize: 13,
-    fontWeight: '600',
-    textAlign: 'center',
-    marginTop: 12,
-    marginBottom: 14,
-  },
-  listRegion: {
-    flex: 1,
-    minHeight: 220,
-  },
-  affirmationListContent: {
-    paddingBottom: 18,
-    paddingTop: 2,
-  },
-  affirmationList: {
-    flex: 1,
-  },
-  affirmationCard: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: 14,
-    borderRadius: 18,
-    paddingHorizontal: 16,
-    paddingVertical: 16,
-    backgroundColor: '#141A26',
-    borderWidth: 1,
-    borderColor: 'rgba(255,255,255,0.06)',
-  },
-  affirmationCardSelected: {
-    borderColor: '#F4C95D',
-    borderWidth: 1.5,
-  },
-  affirmationCardText: {
-    flex: 1,
-    color: '#F5F7FA',
-    fontSize: 16,
-    lineHeight: 23,
-    fontWeight: '500',
-  },
-  addButton: {
-    width: 40,
-    height: 40,
-    borderRadius: 20,
-    backgroundColor: 'rgba(255,255,255,0.08)',
-    alignItems: 'center',
-    justifyContent: 'center',
-    borderWidth: 1,
-    borderColor: 'rgba(255,255,255,0.08)',
-  },
-  addButtonSelected: {
-    backgroundColor: 'rgba(139, 92, 246, 0.55)',
-    borderColor: 'rgba(167, 139, 250, 0.38)',
-  },
-  addButtonPressed: {
-    opacity: 0.9,
-    transform: [{ scale: 0.96 }],
-  },
-  listSeparator: {
-    height: 12,
-  },
-  emptyTabState: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingHorizontal: 20,
-    paddingVertical: 36,
-  },
-  emptyTabTitle: {
-    color: '#F5F7FA',
-    fontSize: 18,
-    fontWeight: '700',
-    textAlign: 'center',
-  },
-  emptyTabCopy: {
-    color: '#C6CDD8',
-    fontSize: 14,
-    lineHeight: 21,
-    textAlign: 'center',
-    marginTop: 8,
-  },
-  modalFooter: {
-    paddingTop: 16,
-    gap: 14,
-    alignItems: 'center',
-  },
-  queueCountPill: {
-    minWidth: 210,
-    minHeight: 44,
-    borderRadius: 999,
-    alignItems: 'center',
-    justifyContent: 'center',
-    paddingHorizontal: 18,
-    backgroundColor: 'rgba(255,255,255,0.06)',
-    borderWidth: 1,
-    borderColor: 'rgba(255,255,255,0.08)',
-  },
-  queueCountText: {
-    color: '#F5F7FA',
-    fontSize: 15,
-    fontWeight: '700',
   },
 });
